test_project/requirements.txt
pillow==2.2.1
fhurl==0.1.7
smarturls==0.1.2
gunicorn==18.0
six==1.4.1
importd==0.2.3
django-mptt==0.5.2
django-treeadmin==0.4.1
html5lib==1.0b3
django-cms==2.4.2
django-sekizai==0.7
django-debug-toolbar==0.9.4
django-pdb==0.3.2
